@@ -1,13 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 
 import 'bootstrap/dist/css/bootstrap.min.css';
 
-ReactDOM.render(
-    <App/>,
-  document.getElementById('root')
-);
+ReactDOM
+    .createRoot(document.getElementById('root'))
+    .render(
+        <App />
+    );
 
 
