@@ -1,97 +1,39 @@
-import { Component, useState, useEffect } from 'react';
-import { Container } from 'react-bootstrap';
-import './App.css';
-
-
-// class Slider extends Component {
-
-//     constructor(props) {
-//         super(props);
-//         this.state = {
-//             autoplay: false,
-//             slide: 0
-//         }
-//     }
-
-//     changeSlide = (i) => {
-//         this.setState(({ slide }) => ({
-//             slide: slide + i
-//         }))
-//     }
-
-//     toggleAutoplay = () => {
-//         this.setState(({ autoplay }) => ({
-//             autoplay: !autoplay
-//         }))
-//     }
-
-//     render() {
-//         return (
-//             <Container>
-//                 <div className="slider w-50 m-auto">
-//                     <img className="d-block w-100" src="https://www.planetware.com/wpimages/2020/02/france-in-pictures-beautiful-places-to-photograph-eiffel-tower.jpg" alt="slide" />
-//                     <div className="text-center mt-5">Active slide {this.state.slide} <br /> {this.state.autoplay ? 'auto' : null}</div>
-//                     <div className="buttons mt-3">
-//                         <button
-//                             className="btn btn-primary me-2"
-//                             onClick={() => this.changeSlide(-1)}>-1</button>
-//                         <button
-//                             className="btn btn-primary me-2"
-//                             onClick={() => this.changeSlide(1)}>+1</button>
-//                         <button
-//                             className="btn btn-primary me-2"
-//                             onClick={this.toggleAutoplay}>toggle autoplay</button>
-//                     </div>
-//                 </div>
-//             </Container>
-//         )
-//     }
-// }
-
-
-const Slider = (props) => {
-
-    const [slide, setSlide] = useState(0);
-    const [autoPlay, setAutoPlay] = useState(false);
-
-    useEffect(() => {
-        document.title = `Slide: ${slide}`;
-    });
-    
-    function changeSlide(i) {
-        setSlide(slide => slide + i);
-    }
-
-    function toggleAutoplay() {
-        setAutoPlay(autoPlay => !autoPlay);
-    }
-
-
-    return (
-        <Container>
-            <div className="slider w-50 m-auto">
-                <img className="d-block w-100" src="https://www.planetware.com/wpimages/2020/02/france-in-pictures-beautiful-places-to-photograph-eiffel-tower.jpg" alt="slide" />
-                <div className="text-center mt-5">Active slide {slide} <br/> {autoPlay ? 'auto' : null}</div>
-                <div className="buttons mt-3">
-                    <button 
-                        className="btn btn-primary me-2"
-                        onClick={() => changeSlide(-1)}>-1</button>
-                    <button 
-                        className="btn btn-primary me-2"
-                        onClick={() => changeSlide(1)}>+1</button>
-                    <button 
-                        className="btn btn-primary me-2"
-                        onClick={toggleAutoplay}>toggle autoplay</button>
-                </div>
-            </div>
-        </Container>
-    )
-}
-
+import data from './data';
+import { useState, useMemo, useDeferredValue, useTransition } from 'react';
 
 function App() {
+    const [text, setText] = useState('');
+    const [posts, setPosts] = useState(data);
+    //const deferredValue = useDeferredValue(text);
+    const [isPending, startTransition] = useTransition();
+
+    const filteredPosts = useMemo(() => {
+        const searchText = text.toLowerCase();
+        return posts.filter(item => item.name.toLowerCase().includes(searchText));
+    }, [text]);
+
+    const onValueChange = (e) => {
+        startTransition(() =>
+            setText(e.target.value)
+        );
+    }
+
     return (
-        <Slider />
+        <>
+            <input value={text} type='text' onChange={onValueChange} />
+
+            <hr />
+
+            <div>
+                {isPending ? <h4>Loading...</h4> :
+                    filteredPosts.map(post => (
+                        <div key={post._id}>
+                            <h4>{post.name}</h4>
+                        </div>
+                    ))
+                }
+            </div>
+        </>
     );
 }
 
